ci: Remove Core App deployment GitHub Action
- Action was moved to core app repo after sequential builds were introduced in GitHub Actions
- Do some code clean
